Rename Ceres loss function files
