+++ src/colmap/estimators/ceres_loss_function.h
